convert int to qint64 during multiplication
Squashed commit of the following:

commit 49ad79595053771a49b43a7ccdec1af5e92e265e
Author: Karsten Borgwaldt <kb@spambri.de>
Date:   Sun Oct 24 10:57:42 2021 +0200

    remove convert-int branch from codeql scanning

commit 7de8e04c5faec55fac0d2569f431a4e45ec36700
Author: Karsten Borgwaldt <kb@spambri.de>
Date:   Sun Oct 24 10:44:52 2021 +0200

    cast int to qint64 before multiplication

+++ src/timeinputdialog.cpp
@@ -163,5 +163,5 @@ QDateTime TimeSpec::toDateTime() const
 		return m_absolute;
 
 	QDateTime now = QDateTime::currentDateTime();
-	return now.addSecs(-1 * m_relativeValue * m_relativeUnit);
+	return now.addSecs(-1 * static_cast<qint64>(m_relativeValue) * m_relativeUnit);
 }
